serpapi_tools/requirements.txt
streamlit>=1.31.0
pandas>=2.0.0
python-dotenv>=1.0.0
google-search-results>=2.4.2
